Revert "Создал отдельный h и cpp файлы для существующих классов и временно внес их туда. Получилось все еще громоздко. Есть смысл разделить некоторые классы еще на несколько файлов."
